correct bugs in rule based manager

+++ src/applications/model/prc-monitor.h
@@ -63,7 +63,7 @@ private:
 
   void UpdateStats (void);
   
-  Ptr<Packet> CreateTrapPacket(std::string mib, uint32_t value, Mac48Address address);
+  Ptr<Packet> CreateTrapPacket(std::string mib, double value, Mac48Address address);
 
   /**
    * \brief create a Packet with a suscription
