+++ src/wifi/model/rule-based-wifi-manager.h
@@ -59,7 +59,7 @@ public:
   void IncreaseRate (Mac48Address address, uint32_t levels);
   void DecreasePower (Mac48Address address, uint32_t levels);
   void IncreasePower (Mac48Address address, uint32_t levels);
-  uint32_t GetStats (Mac48Address address);
+  double GetStats (Mac48Address address);
   uint32_t GetCurrentRate (Mac48Address address);
   uint8_t GetCurrentPower (Mac48Address address);
